feat(engine.rs): Implemented everything into a wrapper type.
This works. It friggin works. Oh my, I don't care I am taking a break, and I'll go get a cookie. This is not the cleanest, but the implementation works and while I have no visualizer implemented,  by making and deriving a graph on paper, I am confident the code is correct.

Now I need to implement the differentiation on the whole graph, and after that the neural network component.